clean up option-checker
